Extract config/workspace/naming/ssh mixins from BoxmanManager (#85 item 29, part 1)
Move-only refactor: cohesive low-coupling method groups move verbatim
into mixin classes under src/boxman/manager_parts/ (ConfigMixin,
WorkspaceMixin, NamingMixin, SSHMixin); BoxmanManager multiply-inherits
them so every self.* cross-call keeps working unchanged.

Refs #85
