+++ src/boxman/manager.py
@@ -3,7 +3,6 @@ import hashlib
 import json
 import logging
 import os
-import re
 import shutil
 import subprocess
 import time
@@ -19,17 +18,19 @@ from boxman.config_cache import BoxmanCache
 from boxman.exceptions import ConfigError, ProvisionError, SnapshotError
 from boxman.image_cache import ImageCache
 from boxman.loggers.logger import suppressed
+from boxman.manager_parts.config import ConfigMixin
+from boxman.manager_parts.naming import NamingMixin
+from boxman.manager_parts.ssh import SSHMixin
+from boxman.manager_parts.workspace import WorkspaceMixin
 from boxman.netlab import ContainerlabManager, shared_bridges
-from boxman.providers import PROVIDERS, merge_provider_configs, primary_provider_type
+from boxman.providers import merge_provider_configs, primary_provider_type
 from boxman.providers.libvirt import net_reconcile
 from boxman.providers.libvirt.commands import VirshCommand
 from boxman.providers.libvirt.virsh_parse import parse_domblklist
 from boxman.runtime import RuntimeBase, create_runtime
 from boxman.task_runner import TaskRunner
 from boxman.utils.http_download import download_url
-from boxman.utils.io import write_files
 from boxman.utils.jinja_env import create_jinja_env
-from boxman.utils.references import resolve_reference
 from boxman.utils.shell import run
 
 #: Seconds to wait for a finished child's queue message before declaring its
@@ -97,7 +98,7 @@ def _clone_with_retry(provider, cluster, vm_info, new_vm_name,
                 raise
 
 
-class BoxmanManager:
+class BoxmanManager(ConfigMixin, WorkspaceMixin, NamingMixin, SSHMixin):
     def __init__(self,
                  config: dict[str, Any] | None = None):
         """
@@ -152,15 +153,6 @@ class BoxmanManager:
 
         #: Optional[Dict[str, Any]]: the boxman application-level config (from boxman.yml)
         self.app_config: dict[str, Any] | None = None
-
-    def load_app_config(self, config: dict[str, Any]) -> None:
-        """
-        Load the boxman application-level configuration (from boxman.yml).
-
-        Args:
-            config: The parsed boxman.yml configuration dictionary
-        """
-        self.app_config = config
 
     @property
     def provider(self) -> Optional["ProviderSession"]:
@@ -666,961 +658,6 @@ class BoxmanManager:
             A new dict with ``runtime`` and related keys injected.
         """
         return self.runtime_instance.inject_into_provider_config(provider_config)
-
-    def load_config(self, config_path: str) -> dict[str, Any]:
-        """
-        Load configuration from a YAML file.
-
-        The file is first treated as a Jinja template and rendered,
-        then parsed as YAML.
-
-        Args:
-            config_path: Path to the configuration file
-
-        Returns:
-            Dict containing the configuration
-        """
-        # get the directory and filename for jinja template loading
-        config_dir = os.path.dirname(os.path.abspath(config_path))
-        config_filename = os.path.basename(config_path)
-
-        # expose the config file location as env vars so templates can use
-        # {{ env("BOXMAN_CONF_FILE") }} and {{ env("BOXMAN_CONF_DIR") }}.
-        # setdefault lets a user-defined value take precedence.
-        os.environ.setdefault('BOXMAN_CONF_FILE', os.path.abspath(config_path))
-        os.environ.setdefault('BOXMAN_CONF_DIR', config_dir)
-
-        # create jinja environment with boxman helpers (env(), env_required(), etc.)
-        env = create_jinja_env(config_dir)
-
-        # Read raw content and convert bare {{ name }} placeholders to
-        # {name} markers before Jinja2 rendering.  Bare names (no parens,
-        # dots, pipes, etc.) are task-command placeholders, not real Jinja2
-        # expressions — real ones are function calls like {{ env("VAR") }}.
-        #
-        # IMPORTANT: variables defined by Jinja2 control flow ({% for %},
-        # {% set %}) must be excluded from this substitution so that Jinja2
-        # can render them correctly.  Without this exclusion, a loop like
-        #   {% for suffix in 'abc' %} ... disk{{ suffix }} ... {% endfor %}
-        # would have {{ suffix }} converted to {suffix} before Jinja2 runs,
-        # leaving the literal string "disk{suffix}" in the output.
-        raw_path = os.path.join(config_dir, config_filename)
-        with open(raw_path) as fobj:
-            raw_content = fobj.read()
-
-        # collect all variable names introduced by jinja2 control flow
-        jinja_ctrl_vars: set = set()
-        jinja_ctrl_vars.update(re.findall(r'\{%-?\s*for\s+(\w+)\s+in\b', raw_content))
-        jinja_ctrl_vars.update(re.findall(r'\{%-?\s*set\s+(\w+)\s*=', raw_content))
-
-        def _preserve_jinja_vars(m: re.Match) -> str:
-            name = m.group(1)
-            if name in jinja_ctrl_vars:
-                return m.group(0)   # keep {{ name }} for Jinja2 to render
-            return '{' + name + '}'  # task placeholder → {name}
-
-        preserved = re.sub(r"\{\{\s*(\w+)\s*\}\}", _preserve_jinja_vars, raw_content)
-
-        # load and render the template from the pre-processed string
-        template = env.from_string(preserved)
-
-        # render the template
-        # NOTE: pass os.environ as 'environ' (not 'env') to avoid shadowing
-        # the env() helper function registered in the Jinja globals.
-        rendered_yaml = template.render(
-            environ=os.environ,
-        )
-
-        # parse the rendered yaml
-        conf = yaml.safe_load(rendered_yaml)
-
-        # dump the rendered yaml file for debugging/inspection
-        rendered_filename = f"{os.path.splitext(config_filename)[0]}.rendered.yml"
-        rendered_path = os.path.join(config_dir, rendered_filename)
-        with open(rendered_path, 'w') as fobj:
-            fobj.write(rendered_yaml)
-            self.logger.info(f"rendered YAML template written to {rendered_path}")
-
-        # apply schema-version handling (v2.0 boxes:→vms: normalization etc.)
-        # before returning, so every downstream consumer sees the internal
-        # (v1.0-shaped) config regardless of the on-disk schema version.
-        conf = self._apply_config_version(conf)
-
-        return conf
-
-    def _apply_config_version(self, conf: dict[str, Any] | None) -> dict[str, Any] | None:
-        """
-        Dispatch on the config's ``version:`` key and return the config in
-        boxman's internal (v1.0) shape.
-
-        - No ``version:`` key, ``'1.0'`` (or the unquoted YAML numeric
-          ``1`` / ``1.0``) → returned unchanged (v1.0 is supported
-          indefinitely and stays byte-identical).
-        - ``'2.0'`` (or unquoted ``2`` / ``2.0``) → :meth:`normalize_v2_config`
-          (per-cluster ``boxes:``→``vms:`` for libvirt clusters, schema
-          validation).
-        - anything else → :class:`~boxman.exceptions.ConfigError`.
-
-        The value is compared as a string so unquoted YAML numerics work:
-        ``version: 2`` and ``version: 2.0`` both select v2.0. Quoting
-        (``version: '2.0'``) is still recommended — see
-        ``doc/docker-compose-provider/config-schema.md``.
-
-        Args:
-            conf: The parsed project configuration. A non-mapping root
-                (``None``, list, scalar — e.g. an empty file) is returned
-                unchanged rather than raising, so malformed input surfaces
-                later through the normal path, not as an ``AttributeError``
-                here.
-
-        Returns:
-            The version-normalized configuration.
-
-        Raises:
-            ConfigError: If ``version:`` is set to an unsupported value.
-        """
-        if not isinstance(conf, dict):
-            return conf
-
-        version = str(conf.get('version', '1.0')).strip()
-        if version in ('1', '1.0'):
-            self._reject_v1_docker_compose(conf)
-            self._warn_on_v1_boxes(conf)
-            return conf
-        if version in ('2', '2.0'):
-            return self.normalize_v2_config(conf)
-
-        raise ConfigError(
-            f"unsupported config version: '{version}' "
-            f"(supported: '1.0', '2.0')"
-        )
-
-    def _reject_v1_docker_compose(self, conf: dict[str, Any]) -> None:
-        """
-        Reject a v1.0 / versionless config that resolves any cluster to the
-        docker-compose provider.
-
-        The docker-compose provider consumes ``boxes:``, which v1.0 ignores
-        (see :meth:`_warn_on_v1_boxes`). Without this guard the cluster would
-        be picked up by ``_compose_clusters`` and provisioned into live
-        services *despite* the v1 nudge telling the user those boxes are
-        ignored — the warning and the behaviour would contradict. The
-        provider is new in this epic, so no legitimate v1 config uses it;
-        failing fast with a clear message is safe.
-
-        Raises:
-            ConfigError: If a cluster's effective provider (per-cluster
-                ``provider:`` or the primary provider) is ``docker-compose``.
-        """
-        for cluster_name, cluster in (conf.get('clusters') or {}).items():
-            if not isinstance(cluster, dict):
-                continue
-            provider = cluster.get('provider') or primary_provider_type(conf)
-            if provider == 'docker-compose':
-                raise ConfigError(
-                    f"cluster '{cluster_name}' uses the docker-compose "
-                    f"provider, which requires version: '2.0' (docker-compose "
-                    f"clusters consume 'boxes:', ignored under v1.0). Add "
-                    f"\"version: '2.0'\" to the config."
-                )
-
-    def _warn_on_v1_boxes(self, conf: dict[str, Any]) -> None:
-        """
-        Warn when a v1.0 config uses the v2.0-only ``boxes:`` key.
-
-        In v1.0 nothing reads ``boxes:``, so such boxes are silently
-        ignored — including in a partial migration where a cluster carries
-        both ``vms:`` (provisioned) and ``boxes:`` (dropped). The warning
-        fires whenever ``boxes:`` is present, regardless of ``vms:``. It is
-        a log-only nudge toward ``version: '2.0'`` and does not change v1.0
-        provisioning behaviour.
-        """
-        for cluster_name, cluster in (conf.get('clusters') or {}).items():
-            if isinstance(cluster, dict) and 'boxes' in cluster:
-                self.logger.warning(
-                    f"cluster '{cluster_name}' uses 'boxes:' but the config "
-                    f"is v1.0 — 'boxes:' is only recognised under "
-                    f"version: '2.0', so these boxes are ignored. Add "
-                    f"\"version: '2.0'\" or rename 'boxes:' to 'vms:'."
-                )
-
-    def normalize_v2_config(self, conf: dict[str, Any]) -> dict[str, Any]:
-        """
-        Normalize a v2.0 config into boxman's internal (v1.0) shape.
-
-        For each cluster the effective provider is
-        ``cluster.get('provider') or primary_provider_type(conf)`` (the
-        same resolution :meth:`provider_type_for_cluster` uses). Then, per
-        the design compatibility matrix:
-
-        - **libvirt** clusters: ``boxes:`` is renamed to ``vms:`` so the
-          ~35 existing libvirt call sites are untouched; a cluster still
-          using ``vms:`` is accepted with a deprecation warning; declaring
-          both ``boxes:`` and ``vms:`` is ambiguous and rejected.
-        - **docker-compose** clusters: ``boxes:`` is kept as-is (consumed
-          by the Phase 3 ``DockerComposeSession``); ``vms:`` is rejected.
-        - any **other** provider (the legacy ``virtualbox``, or an unknown
-          value / typo) is rejected — v2.0 supports only ``libvirt`` and
-          ``docker-compose`` today, and silently leaving ``boxes:``
-          untouched would provision an empty cluster.
-
-        A per-box ``provider:`` key is a config error (ADR-001): providers
-        are declared at the cluster level only.
-
-        The config is **mutated in place** and also returned; it is meant
-        to run once over a freshly parsed config (``load_config`` re-parses
-        each call). Re-running over an already-normalized dict — whose
-        ``boxes:`` were popped but whose ``version:`` stays ``'2.0'`` —
-        would treat the renamed ``vms:`` as a legacy key and emit a
-        spurious deprecation warning.
-
-        Args:
-            conf: The parsed v2.0 project configuration.
-
-        Returns:
-            The normalized configuration (same object).
-
-        Raises:
-            ConfigError: On an ambiguous, provider-incompatible, or
-                unknown-provider cluster, or a per-box ``provider:``.
-        """
-        for cluster_name, cluster in (conf.get('clusters') or {}).items():
-            if not isinstance(cluster, dict):
-                continue
-
-            provider = cluster.get('provider') or primary_provider_type(conf)
-            has_boxes = 'boxes' in cluster
-            has_vms = 'vms' in cluster
-
-            # ADR-001: a `provider:` on an individual box is a config error;
-            # inspect both the boxes: mapping and the legacy vms: mapping.
-            self._reject_per_box_provider(cluster_name, cluster.get('boxes'))
-            self._reject_per_box_provider(cluster_name, cluster.get('vms'))
-
-            if provider == 'libvirt':
-                if has_boxes and has_vms:
-                    raise ConfigError(
-                        f"cluster '{cluster_name}' declares both 'boxes:' "
-                        f"and 'vms:' — use one (prefer 'boxes:' in v2.0)."
-                    )
-                if has_boxes:
-                    cluster['vms'] = cluster.pop('boxes')
-                elif has_vms:
-                    self.logger.warning(
-                        f"cluster '{cluster_name}' uses the legacy 'vms:' "
-                        f"key under version: '2.0' — 'boxes:' is the "
-                        f"preferred generic key. 'vms:' remains accepted."
-                    )
-            elif provider == 'docker-compose':
-                if has_vms:
-                    raise ConfigError(
-                        f"cluster '{cluster_name}' declares 'vms:'; "
-                        f"docker-compose clusters only support 'boxes:'."
-                    )
-                # 'boxes:' is kept verbatim for the docker-compose provider.
-            elif provider in PROVIDERS:
-                # a registered provider not yet wired for v2.0 (virtualbox)
-                raise ConfigError(
-                    f"cluster '{cluster_name}': provider '{provider}' is "
-                    f"not supported under config version '2.0' yet — use "
-                    f"version: '1.0' with 'vms:' for '{provider}' clusters."
-                )
-            else:
-                raise ConfigError(
-                    f"cluster '{cluster_name}': unknown provider "
-                    f"'{provider}' (known: {', '.join(sorted(PROVIDERS))})."
-                )
-
-        return conf
-
-    @staticmethod
-    def _reject_per_box_provider(cluster_name: str, boxes: Any) -> None:
-        """
-        Raise if any box in *boxes* declares a ``provider:`` key.
-
-        ADR-001 assigns per-box-provider validation to Phase 2: a provider
-        is a cluster-level concern, so a ``provider:`` inside an individual
-        box (which ``provider_type_for_cluster`` would silently ignore) is
-        rejected rather than accepted and dropped.
-        """
-        for box_name, box in (boxes or {}).items():
-            if isinstance(box, dict) and 'provider' in box:
-                raise ConfigError(
-                    f"box '{box_name}' in cluster '{cluster_name}' declares "
-                    f"a 'provider:' key — per-box providers are not "
-                    f"supported (ADR-001); declare 'provider:' on the "
-                    f"cluster instead."
-                )
-
-    @staticmethod
-    def _render_inventory(host_aliases, cluster_groups, host_extra_vars=None) -> str:
-        """
-        Render an Ansible ``01-hosts.yml`` body.
-
-        Args:
-            host_aliases: iterable of ``(host_key, boxman_alias)`` placed under
-                ``all.hosts``.
-            cluster_groups: mapping of group name → list of host keys, rendered
-                as ``all.children.<group>.hosts``.
-            host_extra_vars: optional ``{host_key: {var: value}}`` of extra host
-                vars (e.g. ``ansible_connection``/``ansible_host`` for
-                docker-compose containers reached via ``community.docker``).
-                VM hosts pass nothing and render exactly as before.
-
-        Returns:
-            The YAML text (identical in shape to what boxman has always
-            generated for the combined workspace inventory).
-        """
-        host_extra_vars = host_extra_vars or {}
-        host_blocks: list[str] = []
-        for host, alias in host_aliases:
-            lines = [f'        {host}:', f'          boxman_alias: "{alias}"']
-            for var, value in (host_extra_vars.get(host) or {}).items():
-                lines.append(f'          {var}: "{value}"')
-            host_blocks.append('\n'.join(lines))
-        host_lines = '\n'.join(host_blocks)
-        children_lines: list[str] = []
-        for group, hosts in cluster_groups.items():
-            children_lines.append(f'    {group}:')
-            children_lines.append('      hosts:')
-            for host in hosts:
-                children_lines.append(f'        {host}:')
-        children_section = '\n'.join(children_lines)
-        return (
-            f"---\n"
-            f"all:\n"
-            f"  hosts:\n"
-            f"{host_lines}\n"
-            f"  children:\n"
-            f"{children_section}\n"
-        )
-
-    @staticmethod
-    def _cluster_inventory_key(cluster: dict) -> str:
-        """
-        Resolve the ``files`` key for a cluster's own ``01-hosts.yml``.
-
-        Honors a per-cluster ``inventory:`` override (relative to the cluster
-        workdir, or absolute); otherwise defaults to ``inventory/01-hosts.yml``
-        under the cluster workdir. The key is consumed by ``provision_files``,
-        which writes cluster files with ``rootdir=cluster['workdir']`` — so an
-        absolute override naturally bypasses the workdir, mirroring how the
-        workspace-level custom inventory path is handled.
-        """
-        inv = cluster.get('inventory')
-        if inv:
-            inv = os.path.normpath(os.path.expanduser(str(inv)))
-            return os.path.join(inv, '01-hosts.yml')
-        return os.path.join('inventory', '01-hosts.yml')
-
-    def resolve_workspace_defaults(self) -> None:
-        """
-        Resolve workspace defaults for each cluster.
-
-        For each cluster:
-        - If workdir is not set, default to workspace.path / cluster_name
-        - Auto-generate a per-cluster ``inventory/01-hosts.yml`` (location
-          overridable via the cluster's ``inventory:`` key) containing ONLY
-          that cluster's hosts, so per-cluster Ansible runs are isolated.
-
-        At the workspace level (written to workspace.path):
-        - Auto-generate env.sh, ansible.cfg, and a combined
-          inventory/01-hosts.yml spanning every cluster (used for
-          project-wide ops and ``boxman ssh`` alias resolution).
-        """
-        config = self.config
-        workspace = config.get('workspace', {})
-        workspace_path = workspace.get('path', '')
-        clusters = config.get('clusters', {})
-
-        # workspace-level files (written to workspace.path)
-        ws_files = workspace.setdefault('files', {})
-
-        def _resolve_output_path(custom_path, workspace_path):
-            """Resolve a custom path: expanduser, make absolute relative to workspace.path, normpath."""
-            p = os.path.expanduser(custom_path)
-            if not os.path.isabs(p):
-                p = os.path.join(os.path.expanduser(workspace_path), p)
-            return os.path.normpath(p)
-
-        # Determine output keys for generated files based on workspace config overrides.
-        # When a custom path is set and resolves to an absolute path, os.path.join(rootdir, abs)
-        # returns the absolute path — so it naturally bypasses rootdir in write_files().
-        custom_ansible_config = workspace.get('ansible_config')
-        custom_env_file = workspace.get('env_file')
-        custom_inventory = workspace.get('inventory')
-
-        if custom_env_file and workspace_path:
-            env_sh_key = _resolve_output_path(custom_env_file, workspace_path)
-        else:
-            env_sh_key = 'env.sh'
-
-        if custom_inventory and workspace_path:
-            inv_dir = _resolve_output_path(custom_inventory, workspace_path)
-            inventory_key = os.path.join(inv_dir, '01-hosts.yml')
-        else:
-            inventory_key = 'inventory/01-hosts.yml'
-
-        if custom_ansible_config and workspace_path:
-            ansible_cfg_key = _resolve_output_path(custom_ansible_config, workspace_path)
-        else:
-            ansible_cfg_key = 'ansible.cfg'
-
-        for cluster_name, cluster in clusters.items():
-            # resolve workdir: explicit > workspace.path/cluster_name (for both
-            # libvirt and docker-compose clusters).
-            if 'workdir' not in cluster:
-                if workspace_path:
-                    cluster['workdir'] = os.path.join(workspace_path, cluster_name)
-                else:
-                    self.logger.warning(
-                        f"cluster '{cluster_name}' has no workdir and "
-                        f"workspace.path is not set"
-                    )
-
-        # --- inventory generation ---
-        # Project-wide host ordering, once, so a host's boxman_alias is
-        # identical in the combined and per-cluster inventories (and lines up
-        # with the node<N> ssh_config aliases). Rows are (cluster, name,
-        # host_key, extra_vars): libvirt VMs reach via ssh (no extra vars);
-        # docker-compose containers reach via the community.docker connection
-        # to the deterministic compose container name (<project>-<box>-1).
-        all_hosts: list[tuple[str, str, str, dict]] = []
-        for cname, cluster in clusters.items():
-            if self._is_compose_cluster(cname):
-                project = self._compose_project_for(cname)
-                for box in (cluster.get('boxes') or {}):
-                    all_hosts.append((cname, box, f'{cname}_{box}', {
-                        'ansible_connection': 'community.docker.docker',
-                        'ansible_host': f'{project}-{box}-1',
-                    }))
-            else:
-                for vm_name in (cluster.get('vms') or {}):
-                    all_hosts.append((cname, vm_name, f'{cname}_{vm_name}', {}))
-
-        pad_width = len(str(len(all_hosts) - 1)) if len(all_hosts) > 1 else 1
-        alias_of = {
-            (c, n): f"node{str(i).zfill(pad_width)}"
-            for i, (c, n, _hk, _ev) in enumerate(all_hosts)
-        }
-
-        # --- env.sh (workspace-level) --- GATEWAYHOST is the first libvirt VM
-        # (the `boxman ssh` default target); a dc-only project leaves it empty
-        # since containers are reached with `boxman exec`, not ssh.
-        if all_hosts and env_sh_key not in ws_files:
-            first_vm = next((hk for (_c, _n, hk, ev) in all_hosts if not ev), '')
-            inv_val = custom_inventory if custom_inventory else 'inventory'
-            cfg_val = custom_ansible_config if custom_ansible_config else 'ansible.cfg'
-            ws_files[env_sh_key] = (
-                f"export INVENTORY={inv_val}\n"
-                f"export SSH_CONFIG=ssh_config\n"
-                f"export GATEWAYHOST={first_vm}\n"
-                f"export ANSIBLE_CONFIG={cfg_val}\n"
-                f"export ANSIBLE_INVENTORY=\"$INVENTORY\"\n"
-                f"export ANSIBLE_SSH_ARGS=\"-F $SSH_CONFIG\"\n"
-            )
-
-        # Combined workspace inventory (every cluster's hosts). Kept for
-        # project-wide ops and `boxman ssh <alias>` resolution. NOTE: because
-        # it merges all clusters under all.hosts, an Ansible consumer that
-        # iterates groups['all'] would see every cluster — which is why each
-        # cluster also gets its own scoped inventory below.
-        if all_hosts and inventory_key not in ws_files:
-            host_aliases = [(hk, alias_of[(c, n)]) for (c, n, hk, _ev) in all_hosts]
-            cluster_groups: dict[str, list[str]] = {}
-            host_extra: dict[str, dict] = {}
-            for (c, _n, hk, ev) in all_hosts:
-                cluster_groups.setdefault(c, []).append(hk)
-                if ev:
-                    host_extra[hk] = ev
-            ws_files[inventory_key] = self._render_inventory(
-                host_aliases, cluster_groups, host_extra)
-
-        # Per-cluster inventory (only that cluster's hosts). This is what a
-        # `run --cluster <name>` consumes (load_workspace_env repoints
-        # INVENTORY/ANSIBLE_INVENTORY at the cluster's `inventory:`), so
-        # groups['all'] never spans clusters. Written under the cluster's own
-        # inventory dir: `cluster.inventory` if set, else <workdir>/inventory.
-        ws_path_abs = (os.path.abspath(os.path.expanduser(workspace_path))
-                       if workspace_path else os.path.abspath('.'))
-        combined_inv_abs = os.path.abspath(
-            os.path.join(ws_path_abs, inventory_key))
-        for cluster_name, cluster in clusters.items():
-            cluster_hosts = [
-                (n, hk, ev) for (c, n, hk, ev) in all_hosts if c == cluster_name
-            ]
-            if not cluster_hosts or 'workdir' not in cluster:
-                continue
-            workdir_abs = os.path.abspath(os.path.expanduser(cluster['workdir']))
-            cluster_inv_key = self._cluster_inventory_key(cluster)
-            cluster_inv_abs = os.path.abspath(
-                os.path.join(workdir_abs, cluster_inv_key))
-
-            # Guard: if this cluster's inventory resolves to the same file as
-            # the combined workspace inventory (e.g. cluster.workdir ==
-            # workspace.path), writing it would clobber the combined file that
-            # `boxman ssh <alias>` relies on. Skip it and leave the combined
-            # one intact rather than silently overwriting it.
-            if cluster_inv_abs == combined_inv_abs:
-                self.logger.warning(
-                    f"cluster '{cluster_name}' inventory resolves to the "
-                    f"combined workspace inventory ({combined_inv_abs}); "
-                    f"skipping its per-cluster inventory to avoid overwriting "
-                    f"it. Give the cluster a distinct workdir or `inventory:` "
-                    f"to isolate it.")
-                continue
-
-            # Auto-wire `cluster.inventory` (default <workdir>/inventory) so
-            # `run --cluster <name>` repoints at this tree without the user
-            # having to declare it; load_workspace_env resolves the relative
-            # value against the cluster workdir.
-            cluster.setdefault('inventory', 'inventory')
-
-            cluster_files = cluster.setdefault('files', {})
-            if cluster_inv_key in cluster_files:
-                continue
-            host_aliases = [
-                (hk, alias_of[(cluster_name, n)])
-                for (n, hk, _ev) in cluster_hosts
-            ]
-            host_extra = {hk: ev for (n, hk, ev) in cluster_hosts if ev}
-            cluster_files[cluster_inv_key] = self._render_inventory(
-                host_aliases,
-                {cluster_name: [hk for (_n, hk, _ev) in cluster_hosts]},
-                host_extra,
-            )
-
-        # --- ansible.cfg (workspace-level) ---
-        if ansible_cfg_key not in ws_files:
-            ws_files[ansible_cfg_key] = (
-                "[defaults]\n"
-                "host_key_checking = False\n"
-                "poll_interval = 5\n"
-                "callbacks_enabled = timer\n"
-                "forks = 10\n"
-                "nocows = 1\n"
-                "timeout = 30\n"
-                "interpreter_python = auto_silent\n"
-                "gathering = smart\n"
-                "fact_caching = jsonfile\n"
-                f"fact_caching_connection = {workspace_path}/.ansible_facts\n"
-                "fact_caching_timeout = 86400\n"
-                "ansible_managed = Ansible managed: {file} modified on "
-                "%Y-%m-%d %H:%M:%S by {uid} on {host}\n"
-                "\n"
-                "[ssh_connection]\n"
-                "pipelining = True\n"
-                "ssh_args = -o ControlMaster=auto -o ControlPersist=60s\n"
-                "control_path = /tmp/ansible-ssh-%%h-%%p-%%r\n"
-            )
-
-    def collect_workdirs(self) -> list:
-        """
-        Return the absolute paths of every workdir referenced by the project
-        config: ``workspace.path`` (if set), one per cluster, plus one per
-        template (falling back to the shared default ``~/boxman-templates``
-        when a template doesn't set its own ``workdir``).
-
-        The list is de-duplicated and path-expanded; callers use it to
-        bind-mount host directories into the runtime container and to
-        enforce runtime/workdir sentinel checks.
-        """
-        dirs: set = set()
-        config = self.config or {}
-
-        ws_path = (config.get('workspace') or {}).get('path')
-        if ws_path:
-            dirs.add(os.path.abspath(os.path.expanduser(ws_path)))
-
-        for cluster in (config.get('clusters') or {}).values():
-            wd = cluster.get('workdir')
-            if wd:
-                dirs.add(os.path.abspath(os.path.expanduser(wd)))
-
-        templates = config.get('templates') or {}
-        if templates:
-            default_tpl_wd = os.path.abspath(
-                os.path.expanduser('~/boxman-templates'))
-            for tpl in templates.values():
-                wd = tpl.get('workdir')
-                if wd:
-                    dirs.add(os.path.abspath(os.path.expanduser(wd)))
-                else:
-                    dirs.add(default_tpl_wd)
-
-        return sorted(dirs)
-
-    #: Filename of the marker written inside every workdir to record which
-    #: runtime most recently provisioned it. Used to detect cross-runtime
-    #: reuse (e.g. a workdir first used with 'local' then re-used with
-    #: 'docker-compose') and prompt the user to switch to a separate path.
-    RUNTIME_SENTINEL_FILENAME = '.boxman-runtime'
-
-    @staticmethod
-    def _canonical_runtime_name(name: str | None) -> str | None:
-        """
-        Normalise runtime names so ``docker``, ``docker-compose`` and
-        any historical variant collapse to a single canonical form used
-        for sentinel comparisons. Without this, a sentinel written as
-        ``docker`` (the CLI flag) would be seen as mismatching against
-        the internal runtime instance name ``docker-compose``.
-        """
-        if not name:
-            return name
-        n = name.strip().lower()
-        if n in ('docker', 'docker-compose'):
-            return 'docker-compose'
-        return n
-
-    def _read_runtime_sentinel(self, path: str) -> str | None:
-        """Return the runtime name recorded in *path*'s sentinel, or None."""
-        sentinel = os.path.join(path, self.RUNTIME_SENTINEL_FILENAME)
-        try:
-            with open(sentinel) as fobj:
-                return fobj.read().strip() or None
-        except OSError:
-            return None
-
-    def _write_runtime_sentinel(self, path: str, runtime: str) -> None:
-        """
-        Record that *path* is owned by *runtime* by writing a
-        ``.boxman-runtime`` marker. Failures are non-fatal: the worst case
-        is that the user sees the collision prompt again on the next run.
-        """
-        if not path or not os.path.isdir(path):
-            return
-        sentinel = os.path.join(path, self.RUNTIME_SENTINEL_FILENAME)
-        try:
-            with open(sentinel, 'w') as fobj:
-                fobj.write(f"{runtime.strip()}\n")
-        except OSError as exc:
-            self.logger.debug(f"could not write {sentinel}: {exc}")
-
-    @staticmethod
-    def _runtime_suffix_for(runtime: str) -> str:
-        """Return the suffix appended to a suggested path, e.g.
-        'docker-runtime' for a docker-compose runtime."""
-        base = runtime.split('-')[0]  # 'docker-compose' -> 'docker'
-        return f"{base}-runtime"
-
-    def _prompt_workdir_runtime_collision(
-        self, path: str, runtime: str
-    ) -> str:
-        """
-        If *path* already exists on disk and was last used by a different
-        runtime, interactively prompt the user to switch to a
-        runtime-specific alternative like ``<path>-docker-runtime``.
-
-        Returns the path to use for this session (either *path* unchanged
-        or the suggested alternative). The user's choice is not persisted
-        to ``conf.yml``; a hint is logged so they can update it manually.
-        """
-        if not path:
-            return path
-        expanded = os.path.abspath(os.path.expanduser(path))
-
-        if not os.path.isdir(expanded):
-            return path
-
-        # a dir that contains only our sentinel (or is empty) cannot
-        # conflict — skip the prompt.
-        try:
-            entries = [
-                e for e in os.listdir(expanded)
-                if e != self.RUNTIME_SENTINEL_FILENAME
-            ]
-        except OSError:
-            return path
-
-        previous = self._read_runtime_sentinel(expanded)
-        if (self._canonical_runtime_name(previous)
-                == self._canonical_runtime_name(runtime)):
-            return path
-        if not entries and previous is None:
-            return path
-
-        suggested = f"{expanded.rstrip(os.sep)}-" \
-                    f"{self._runtime_suffix_for(runtime)}"
-
-        print()
-        print(
-            f"workdir '{expanded}' was last used with runtime "
-            f"'{previous or 'unknown'}' (current runtime: '{runtime}')."
-        )
-        print(
-            "Mixing artifacts across runtimes can cause confusing "
-            "failures (paths baked into ansible.cfg, ssh_config, etc.)."
-        )
-        try:
-            answer = input(
-                f"Use a runtime-specific path "
-                f"'{suggested}' for this run? [y/N]: "
-            ).strip().lower()
-        except EOFError:
-            answer = ''
-
-        if answer in ('y', 'yes'):
-            self.logger.info(
-                f"hint: update conf.yml to use '{suggested}' to persist "
-                f"this choice across runs"
-            )
-            return suggested
-
-        self.logger.warning(
-            f"continuing with '{expanded}' despite runtime mismatch "
-            f"(previous: '{previous or 'unknown'}', current: '{runtime}')"
-        )
-        return path
-
-    def reconcile_workdirs_with_runtime(self, runtime_name: str) -> None:
-        """
-        For each workdir referenced by the project config
-        (``workspace.path``, per-cluster ``workdir``, per-template
-        ``workdir``), detect cross-runtime collisions and rewrite the
-        path in-memory when the user accepts a suffixed alternative.
-
-        No-op for the ``local`` runtime — it is the historical default
-        and doesn't need to be quarantined away from a user's existing
-        layouts.
-        """
-        if runtime_name == 'local':
-            return
-
-        config = self.config or {}
-        workspace = config.setdefault('workspace', {})
-        old_ws_path = workspace.get('path')
-
-        if old_ws_path:
-            old_abs = os.path.abspath(os.path.expanduser(old_ws_path))
-            new_path = self._prompt_workdir_runtime_collision(
-                old_abs, runtime_name)
-            new_abs = os.path.abspath(os.path.expanduser(new_path))
-            if new_abs != old_abs:
-                workspace['path'] = new_abs
-                # Migrate cluster workdirs that live underneath the old
-                # workspace path so they track the rewritten root.
-                for cluster in (config.get('clusters') or {}).values():
-                    wd = cluster.get('workdir')
-                    if not wd:
-                        continue
-                    wd_abs = os.path.abspath(os.path.expanduser(wd))
-                    if wd_abs == old_abs:
-                        cluster['workdir'] = new_abs
-                    elif wd_abs.startswith(old_abs + os.sep):
-                        cluster['workdir'] = (
-                            new_abs + wd_abs[len(old_abs):]
-                        )
-
-        # Clusters whose workdir is explicitly set outside of
-        # workspace.path still need an individual collision check.
-        for cluster in (config.get('clusters') or {}).values():
-            wd = cluster.get('workdir')
-            if not wd:
-                continue
-            wd_abs = os.path.abspath(os.path.expanduser(wd))
-            new_path = self._prompt_workdir_runtime_collision(
-                wd_abs, runtime_name)
-            new_abs = os.path.abspath(os.path.expanduser(new_path))
-            if new_abs != wd_abs:
-                cluster['workdir'] = new_abs
-
-        for tpl in (config.get('templates') or {}).values():
-            wd = tpl.get('workdir') or '~/boxman-templates'
-            wd_abs = os.path.abspath(os.path.expanduser(wd))
-            new_path = self._prompt_workdir_runtime_collision(
-                wd_abs, runtime_name)
-            new_abs = os.path.abspath(os.path.expanduser(new_path))
-            if new_abs != wd_abs:
-                tpl['workdir'] = new_abs
-
-    def provision_files(self) -> None:
-        """
-        Provision files specified in the cluster and workspace configuration.
-        """
-        # workspace-level files (e.g. env.sh) → written to workspace.path
-        workspace = self.config.get('workspace', {})
-        workspace_path = workspace.get('path', '')
-        if workspace_path:
-            if ws_files := workspace.get('files'):
-                write_files(ws_files, rootdir=workspace_path)
-            self._write_runtime_sentinel(
-                os.path.abspath(os.path.expanduser(workspace_path)),
-                self._runtime_name,
-            )
-
-        # cluster-level files → written to cluster workdir
-        clusters = self.config['clusters']
-        for _cluster_name, cluster in clusters.items():
-            if files := cluster.get('files'):
-                write_files(files, rootdir=cluster['workdir'])
-            wd = cluster.get('workdir')
-            if wd:
-                self._write_runtime_sentinel(
-                    os.path.abspath(os.path.expanduser(wd)),
-                    self._runtime_name,
-                )
-
-    def deprovision_files(self) -> None:
-        """
-        Remove files and directories created during provisioning.
-
-        This includes:
-        - Files listed under workspace.files and cluster.files
-        - Generated SSH keys and ssh_config
-        - Cluster workdirs (if empty after cleanup)
-        """
-        workspace = self.config.get('workspace', {})
-        workspace_path = workspace.get('path', '')
-
-        # Remove files listed in workspace.files
-        if workspace_path:
-            if ws_files := workspace.get('files'):
-                self._remove_files(ws_files, rootdir=workspace_path)
-
-        clusters = self.config['clusters']
-        for _cluster_name, cluster in clusters.items():
-            base_path = workspace_path or cluster.get('workdir', '')
-            base_path = os.path.expanduser(base_path)
-
-            # Remove files listed in cluster.files
-            if files := cluster.get('files'):
-                self._remove_files(files, rootdir=cluster['workdir'])
-
-            # Remove generated SSH keys
-            admin_key_name = cluster.get('admin_key_name', 'id_ed25519_boxman')
-            ssh_key_files = {
-                admin_key_name: "",
-                f"{admin_key_name}.pub": "",
-            }
-            self._remove_files(ssh_key_files, rootdir=base_path)
-
-            # Remove generated ssh_config
-            ssh_config_name = cluster.get('ssh_config', 'ssh_config')
-            self._remove_files({ssh_config_name: ""}, rootdir=base_path)
-
-            # Remove cluster workdir if empty
-            cluster_workdir = os.path.expanduser(cluster.get('workdir', ''))
-            if cluster_workdir and os.path.isdir(cluster_workdir):
-                try:
-                    os.rmdir(cluster_workdir)
-                    log.info(f'removed empty directory {cluster_workdir}')
-                except OSError:
-                    pass
-
-    @staticmethod
-    def _remove_files(files: dict[str, str], rootdir: str = None) -> None:
-        rootdir_abs = os.path.normpath(os.path.expanduser(rootdir)) if rootdir else None
-        candidate_dirs: set = set()
-        for _fpath in files:
-            if rootdir:
-                fpath = os.path.join(rootdir, _fpath)
-            else:
-                fpath = _fpath
-            fpath = os.path.normpath(os.path.expanduser(fpath))
-            if os.path.exists(fpath):
-                os.remove(fpath)
-                log.info(f'removed file {fpath}')
-            # Collect all ancestor directories up to (but excluding) rootdir,
-            # even if the file was already gone — the dirs may still be empty.
-            dirpath = os.path.dirname(fpath)
-            while dirpath and dirpath != os.path.sep:
-                if rootdir_abs and dirpath == rootdir_abs:
-                    break
-                candidate_dirs.add(dirpath)
-                dirpath = os.path.dirname(dirpath)
-
-        # Remove empty directories (deepest first)
-        for dirpath in sorted(candidate_dirs, key=len, reverse=True):
-            try:
-                os.rmdir(dirpath)
-                log.info(f'removed empty directory {dirpath}')
-            except OSError:
-                pass
-
-    @classmethod
-    def full_network_name(cls,
-                          project_config: dict[str, Any],
-                          cluster_name: str = None,
-                          network_name: str = None) -> str:
-        """
-        Return the computed network name based on how it is resolved in the name string.
-
-        The network name is expected to have the following format:
-
-            <project_name>::<cluster_name>::<base_network_name>
-
-        The project prefix is always added to the network name.
-
-           prj_name = f'bprj__{project_config["project"]}__bprj'
-
-        The cluster name is added after the project name as such
-
-           cluster_name = f'clstr__{cluster_name}__clstr'
-
-        Finally the network name is added as such
-
-            full_network_name = f'{prj_name}__{cluster_name}__{base_network_name}'
-
-        Args:
-            project_config (Dict[str, Any]): The project configuration dictionary.
-            network_name (str): The network name string, e.g. 'base_name',
-              'cluster_name::base_name', or 'project_name::cluster_name::base_name'.
-
-        Returns:
-            str: The fully qualified network name.
-        """
-        parts = network_name.split("::")
-
-        if len(parts) == 3:           # project, cluster, base
-            _project, _cluster_name, _base_name = parts
-            retval = f'bprj__{_project}__bprj'
-            retval = retval + f'__clstr__{_cluster_name}__clstr'
-            retval = retval + f'__{_base_name}'
-            return retval
-        elif len(parts) == 2:         # cluster, base
-            _cluster_name, _base_name = parts
-            retval = f'bprj__{project_config["project"]}__bprj'
-            retval = retval + f'__clstr__{_cluster_name}__clstr'
-            retval = retval + f'__{_base_name}'
-            return retval
-        elif len(parts) == 1:         # base only
-            _base_name = parts[0]
-            retval = f'bprj__{project_config["project"]}__bprj'
-            retval = retval + f'__clstr__{cluster_name}__clstr'
-            retval = retval + f'__{_base_name}'
-            return retval
-        else:
-            raise ValueError(f"Invalid network name format: {network_name}")
-
-    def resolve_adapter_network(self,
-                                adapter: dict[str, Any],
-                                cluster_name: str) -> None:
-        """Mutate *adapter* in place so ``network_source`` is fully qualified.
-
-        Resolution precedence:
-
-        1. If ``network_source`` names an entry in top-level
-           ``shared_networks:``, rewrite to the host Linux bridge name
-           and set ``source_type: 'bridge'``.
-        2. Else if ``is_global`` is true on the adapter, leave as-is.
-        3. Else apply cluster/project namespacing via
-           :meth:`full_network_name`.
-        """
-        name = adapter['network_source']
-        shared = (self.config or {}).get('shared_networks')
-        if shared_bridges.is_shared_bridge(name, shared):
-            adapter['network_source'] = shared_bridges.resolve_bridge(name, shared)
-            adapter['source_type'] = 'bridge'
-            return
-
-        if adapter.get('is_global', False):
-            return
-
-        adapter['network_source'] = self.full_network_name(
-            project_config=self.config,
-            cluster_name=cluster_name,
-            network_name=name,
-        )
 
     def import_image(self, cli_args) -> None:
         """
@@ -3501,575 +2538,6 @@ class BoxmanManager:
             for vm_name, vm_info in cluster['vms'].items()
         ]
         self._run_parallel(processes, op_label='configure and start vm')
-
-    def get_connect_info(self) -> bool:
-        """
-        Gather connection information for all VMs in all clusters.
-
-        Queries all VMs in parallel and returns True only if every VM has at
-        least one IP address.
-
-        Returns:
-            True if all VMs have at least one IP address, False otherwise
-        """
-        prj_name = f'bprj__{self.config["project"]}__bprj'
-        vm_names = [
-            f"{prj_name}_{cluster_name}_{vm_name}"
-            for cluster_name, cluster in self._vm_clusters.items()
-            for vm_name in cluster['vms']
-        ]
-
-        def _check(full_vm_name):
-            return self.session_for_vm(full_vm_name).get_vm_ip_addresses(full_vm_name)
-
-        # Routed through _run_parallel: a crashed child can no longer
-        # deadlock the parent on a blocking result_queue.get().
-        results, _failures = self._run_parallel(
-            [(n, _check, (n,)) for n in vm_names],
-            op_label='ip address check')
-
-        all_vms_have_ip = True
-        for full_vm_name in vm_names:
-            ips = results.get(full_vm_name)
-            if not ips:
-                all_vms_have_ip = False
-                self.logger.warning(f"vm {full_vm_name} does not have an ip address yet")
-
-        return all_vms_have_ip
-
-    def connect_info(self) -> None:
-        """
-        Display connection information for all VMs in all clusters.
-
-        This method displays the VM names, hostnames, IP addresses, and
-        other connection details for all configured VMs.
-        """
-        self.logger.status("=== vm connection information ===")
-        ws_path = self.config.get('workspace', {}).get('path', '')
-
-        prj_name = f'bprj__{self.config["project"]}__bprj'
-        for cluster_name, cluster in self._vm_clusters.items():
-            self.logger.status(f"cluster: {cluster_name}")
-
-            for vm_name, vm_info in cluster['vms'].items():
-                full_vm_name = f"{prj_name}_{cluster_name}_{vm_name}"
-                hostname = vm_info.get('hostname', vm_name)
-
-                self.logger.status(f"vm: {vm_name} (hostname: {hostname})")
-
-                # get the ip addresses for all interfaces
-                ip_addresses = self.session_for_cluster(cluster_name).get_vm_ip_addresses(full_vm_name)
-
-                if ip_addresses:
-                    self.logger.status("  ip addresses:")
-                    for iface, ip in ip_addresses.items():
-                        self.logger.status(f"    {iface}: {ip}")
-                else:
-                    self.logger.status("  ip addresses: not available")
-
-                # get ssh connection information
-                admin_user = cluster.get('admin_user', '<placeholder>')
-                base_path = ws_path or cluster.get('workdir', '~')
-                admin_key = os.path.expanduser(os.path.join(
-                    base_path,
-                    cluster.get('admin_key_name', 'id_ed25519_boxman')
-                ))
-
-                self.logger.status("  connect via ssh:")
-                # show direct connection if ip is available
-                if ip_addresses:
-                    first_ip = next(iter(ip_addresses.values()))
-                    self.logger.status(f"    direct: ssh -i {admin_key} {admin_user}@{first_ip}")
-
-                # show connection using ssh_config if available
-                if 'ssh_config' in cluster:
-                    ssh_config = os.path.expanduser(os.path.join(
-                        base_path,
-                        cluster.get('ssh_config', 'ssh_config')
-                    ))
-                    self.logger.status(f"    via config: ssh -F {ssh_config} {cluster_name}_{hostname}")
-
-                self.logger.status("")
-
-            self.logger.status("")
-
-        # docker-compose clusters: container status + published ports + the
-        # exec entry point (containers are reached with `boxman exec`, not ssh).
-        for cluster_name, cluster in self._compose_clusters.items():
-            self.logger.info(f"cluster: {cluster_name} (docker-compose)")
-            self.logger.info("-" * 60)
-            try:
-                status = {
-                    r["service"]: r for r in
-                    self._dc_session(cluster_name).container_status(
-                        cluster_name, cluster)
-                }
-            except Exception as exc:
-                self.logger.warning(f"  could not query containers: {exc}")
-                status = {}
-            for box_name in (cluster.get("boxes") or {}):
-                row = status.get(box_name, {})
-                state = row.get("state", "not created")
-                health = f" ({row['health']})" if row.get("health") else ""
-                self.logger.info(f"container: {box_name}  [{state}{health}]")
-                if row.get("ports"):
-                    self.logger.info(f"  published ports: {row['ports']}")
-                self.logger.info(f"  connect: boxman exec {cluster_name}.{box_name}")
-                self.logger.info("")
-            self.logger.info("")
-
-    #: Alias of the ProxyJump stanza written into ``ssh_config`` when the
-    #: docker runtime is active. Kept as a class constant so tests and
-    #: downstream tooling can grep for it reliably.
-    SSH_JUMP_HOST_ALIAS = "boxman-libvirt-jump"
-
-    def _docker_ssh_jump_stanza(self) -> str | None:
-        """
-        Return the ``Host boxman-libvirt-jump`` block to prepend to
-        ``ssh_config`` when the docker runtime is active, or ``None``
-        when the runtime is ``local`` (VMs are directly reachable from
-        the host in that case).
-        """
-        from boxman.runtime.docker_compose import DockerComposeRuntime
-
-        if not isinstance(self.runtime_instance, DockerComposeRuntime):
-            return None
-
-        rt = self.runtime_instance
-        return (
-            f"Host {self.SSH_JUMP_HOST_ALIAS}\n"
-            f"    HostName     127.0.0.1\n"
-            f"    Port         {rt.ssh_port}\n"
-            f"    User         qemu_user\n"
-            f"    IdentityFile {rt.ssh_identity_path}\n"
-            f"    StrictHostKeyChecking no\n"
-            f"    UserKnownHostsFile /dev/null\n"
-            f"\n\n"
-        )
-
-    def write_ssh_config(self) -> None:
-        """
-        Generate SSH configuration file for easy access to VMs.
-
-        Creates an SSH config file in the workspace directory that allows
-        simplified access to VMs without typing full connection details.
-
-        Under the docker runtime, a ``Host boxman-libvirt-jump`` stanza
-        is prepended and each VM block gets a ``ProxyJump`` directive so
-        that host-side ``ssh``, ``scp`` and ansible transparently hop
-        through the libvirt container to reach VMs on libvirt's internal
-        NAT network.
-        """
-        ws_path = self.config.get('workspace', {}).get('path', '')
-        prj_name = f'bprj__{self.config["project"]}__bprj'
-
-        # count total VMs across all clusters for zero-padded alias numbering
-        total_vms = sum(
-            len(cluster.get('vms', {}))
-            for cluster in self.config['clusters'].values()
-        )
-        pad_width = len(str(total_vms - 1)) if total_vms > 1 else 1
-        vm_counter = 0
-
-        jump_stanza = self._docker_ssh_jump_stanza()
-
-        # Group clusters by their resolved ssh_config path. When
-        # workspace.path is set every cluster resolves to the SAME path,
-        # so we must write each unique path once with all relevant VM
-        # blocks — opening 'w' once per cluster would have the later
-        # iteration truncate the earlier one's entries.
-        groups: dict[str, list[tuple[str, dict, str]]] = {}
-        for cluster_name, cluster in self._vm_clusters.items():
-            base_path = ws_path or cluster.get('workdir', '~')
-            ssh_config = os.path.expanduser(os.path.join(
-                base_path,
-                cluster.get('ssh_config', 'ssh_config')
-            ))
-            groups.setdefault(ssh_config, []).append(
-                (cluster_name, cluster, base_path)
-            )
-
-        for ssh_config, clusters_in_group in groups.items():
-            self.logger.info(f"writing ssh config to {ssh_config}")
-
-            with open(ssh_config, 'w') as fobj:
-                # write global SSH options
-                fobj.write('Host *\n')
-                fobj.write('    StrictHostKeyChecking no\n')
-                fobj.write('    UserKnownHostsFile /dev/null\n')
-                fobj.write('\n\n')
-
-                # docker runtime: jump host stanza
-                if jump_stanza:
-                    fobj.write(jump_stanza)
-
-                for cluster_name, cluster, base_path in clusters_in_group:
-                    admin_priv_key = os.path.expanduser(os.path.join(
-                        base_path,
-                        cluster.get('admin_key_name', 'id_ed25519_boxman')
-                    ))
-
-                    for vm_name, vm_info in cluster['vms'].items():
-                        full_vm_name = f"{prj_name}_{cluster_name}_{vm_name}"
-                        hostname = vm_info.get('hostname', vm_name)
-                        prefixed_host = f"{cluster_name}_{hostname}"
-                        padded_alias = f"node{str(vm_counter).zfill(pad_width)}"
-                        vm_counter += 1
-
-                        # get the first ip address if available
-                        ip_addresses = self.session_for_cluster(cluster_name).get_vm_ip_addresses(full_vm_name)
-
-                        if ip_addresses:
-                            first_ip = next(iter(ip_addresses.values()))
-
-                            fobj.write(f'Host {prefixed_host} {padded_alias}\n')
-                            fobj.write(f'    Hostname {first_ip}\n')
-                            fobj.write(f'    User {cluster.get("admin_user", "admin")}\n')
-                            fobj.write(f'    IdentityFile {admin_priv_key}\n')
-                            if jump_stanza:
-                                fobj.write(
-                                    f'    ProxyJump {self.SSH_JUMP_HOST_ALIAS}\n')
-                            fobj.write('\n\n')
-                        else:
-                            self.logger.warning(
-                                f"no ip address available for the vm {vm_name}, "
-                                "skipping SSH config entry")
-
-            self.logger.info(f"ssh config file written to {ssh_config}")
-            self.logger.info(f"to connect: ssh -F {ssh_config} <hostname>")
-
-    def generate_ssh_keys(self) -> bool:
-        """
-        Generate SSH keys for connecting to VMs.
-
-        Creates an SSH key pair in the workspace directory if it doesn't
-        already exist.
-
-        Returns:
-            bool: True if successful, False otherwise
-        """
-        success = True
-        ws_path = self.config.get('workspace', {}).get('path', '')
-
-        for _, cluster in self._vm_clusters.items():
-            base_path = os.path.expanduser(ws_path or cluster['workdir'])
-            admin_key_name = cluster.get('admin_key_name', 'id_ed25519_boxman')
-
-            admin_priv_key = os.path.join(base_path, admin_key_name)
-            admin_pub_key = os.path.join(base_path, f"{admin_key_name}.pub")
-
-            # create directory if it doesn't exist
-            if not os.path.isdir(base_path):
-                os.makedirs(base_path, exist_ok=True)
-
-            # generate key pair if it doesn't exist
-            if not os.path.exists(admin_priv_key):
-                self.logger.info(f"generating ssh key pair in {base_path}")
-
-                try:
-                    cmd = f'ssh-keygen -t ed25519 -a 100 -f {admin_priv_key} -q -N ""'
-                    run(cmd, hide=True, warn=True)
-
-                    # verify keys were created
-                    if os.path.isfile(admin_priv_key) and os.path.isfile(admin_pub_key):
-                        self.logger.info(f"ssh key pair successfully generated at {admin_priv_key}")
-                    else:
-                        self.logger.warning(f"failed to generate ssh key pair at {admin_priv_key}")
-                        success = False
-
-                except Exception as exc:
-                    self.logger.error(f"error generating ssh key pair: {exc}")
-                    success = False
-            else:
-                self.logger.info(f"using existing ssh key pair at {admin_priv_key}")
-
-        return success
-
-    def get_global_authorized_keys(self) -> list[str]:
-        """
-        Resolve and return all global SSH authorized keys from the app config.
-
-        Reads ``ssh.authorized_keys`` from :pyattr:`app_config` (the top-level
-        ``boxman.yml``), resolves each entry via :pyfunc:`fetch_value`, and
-        returns a list of public-key strings.
-
-        Returns:
-            List of resolved SSH public key strings.
-        """
-        raw_keys = (
-            (self.app_config or {})
-            .get("ssh", {})
-            .get("authorized_keys", [])
-        )
-        resolved: list[str] = []
-        for entry in raw_keys:
-            try:
-                resolved.append(self.fetch_value(entry))
-            except (ValueError, FileNotFoundError) as exc:
-                self.logger.warning(f"skipping unresolvable SSH key entry: {exc}")
-        return resolved
-
-    def write_global_authorized_keys_file(self, output_path: str) -> None:
-        """
-        Resolve global SSH keys from app_config and write them to a file.
-
-        This bridges the Python-side boxman.yml config with the container
-        entrypoint, which cannot read boxman.yml directly. The entrypoint
-        reads ``global_authorized_keys`` from the bind-mounted ssh dir.
-
-        Args:
-            output_path: Path to write the authorized keys file
-                         (e.g. ``<data_dir>/ssh/global_authorized_keys``).
-        """
-        keys = self.get_global_authorized_keys()
-        if not keys:
-            self.logger.info("no global authorized keys to write")
-            return
-
-        os.makedirs(os.path.dirname(output_path), exist_ok=True)
-        with open(output_path, "w") as fobj:
-            for key in keys:
-                fobj.write(key + "\n")
-        self.logger.info(
-            f"wrote {len(keys)} global authorized key(s) to {output_path}")
-
-    @classmethod
-    def fetch_value(cls, value) -> str:
-        """
-        Resolve a config reference (``${env:VAR}`` / ``file://…``) to its
-        concrete value, or return the literal as-is.
-
-        Thin wrapper around :func:`boxman.utils.references.resolve_reference`
-        — kept on the class so external callers importing
-        ``BoxmanManager.fetch_value`` keep working (Phase 2.4 extraction).
-
-        Raises:
-            ValueError: If the environment variable is not set.
-            FileNotFoundError: If the referenced file does not exist.
-        """
-        return resolve_reference(value)
-
-    def add_ssh_keys_to_vms(self) -> bool:
-        """
-        Add the generated SSH public key to all VMs to enable password-less login.
-
-        Uses sshpass to add the public key to each VM using the admin password.
-
-        Returns:
-            bool: True if all VMs received the key successfully, False otherwise
-        """
-        all_successful = True
-        ws_path = self.config.get('workspace', {}).get('path', '')
-
-        prj_name = f'bprj__{self.config["project"]}__bprj'
-        for cluster_name, cluster in self._vm_clusters.items():
-            base_path = os.path.expanduser(ws_path or cluster['workdir'])
-            admin_key_name = cluster.get('admin_key_name', 'id_ed25519_boxman')
-            admin_pub_key = os.path.join(base_path, f"{admin_key_name}.pub")
-
-            admin_user = cluster.get('admin_user', 'admin')
-            admin_pass = self.fetch_value(cluster.get('admin_pass', None))
-
-            if not admin_pass:
-                self.logger.info(
-                    f"warning: No admin password provided for cluster {cluster_name}, "
-                    "cannot add SSH keys")
-                all_successful = False
-                continue
-
-            if not os.path.isfile(admin_pub_key):
-                self.logger.error(f"error: SSH public key {admin_pub_key} does not exist")
-                all_successful = False
-                continue
-
-            self.logger.info(f"adding ssh public key to VMs in cluster {cluster_name}")
-
-            for vm_name, vm_info in cluster['vms'].items():
-                full_vm_name = f"{prj_name}_{cluster_name}_{vm_name}"
-
-                # get the ip addresses for this vm
-                ip_addresses = self.session_for_cluster(cluster_name).get_vm_ip_addresses(full_vm_name)
-
-                if not ip_addresses:
-                    self.logger.warning(
-                        f"no ip address available for vm {vm_name}, "
-                        "cannot add ssh key")
-                    all_successful = False
-                    continue
-
-                # use first available ip address
-                ip_address = next(iter(ip_addresses.values()))
-
-                self.logger.info(f"adding ssh key to vm {vm_name} ({ip_address})...")
-
-                # try to add the key with exponential backoff
-                hostname = vm_info.get('hostname', vm_name)
-                prefixed_host = f"{cluster_name}_{hostname}"
-                success = self._try_add_ssh_key(
-                    ip_address=ip_address,
-                    hostname=prefixed_host,
-                    admin_user=admin_user,
-                    admin_pass=admin_pass,
-                    pub_key_path=admin_pub_key,
-                    ssh_conf_path=os.path.join(base_path, cluster['ssh_config'])
-                )
-
-                if success:
-                    self.logger.info(f"successfully added the ssh key to the vm {vm_name}")
-                else:
-                    self.logger.error(f"failed to add the ssh key to the vm {vm_name}")
-                    all_successful = False
-
-        return all_successful
-
-    def _try_add_ssh_key(self,
-                         ip_address: str,
-                         hostname: str,
-                         admin_user: str,
-                         admin_pass: str,
-                         pub_key_path: str,
-                         ssh_conf_path: str) -> bool:
-        """
-        Try to add an SSH key to a VM with exponential backoff.
-
-        Args:
-            ip_address: IP address of the VM
-            hostname: Hostname of the VM
-            admin_user: Username for SSH login
-            admin_pass: Password for SSH login
-            pub_key_path: Path to the public key file
-            ssh_conf_path: Path to the SSH config file
-
-        Returns:
-            bool: True if successful, False otherwise
-        """
-        wait_time = 1  # Start with 1 second
-        max_retries = 10
-        max_wait = 60  # Maximum wait per attempt
-
-        for attempt in range(1, max_retries + 1):
-            self.logger.info(
-                f"attempt {attempt}/{max_retries} to add ssh key (waiting {wait_time}s)")
-
-            # use sshpass to add the public key
-            cmd = (
-                f'sshpass -p {admin_pass} ssh-copy-id -i {pub_key_path} '
-                f'-o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null '
-                f'{admin_user}@{ip_address}'
-            )
-
-            # When using a non-local runtime, the VM is only reachable from
-            # inside the container, so wrap the command with docker exec.
-            cmd = self.runtime_instance.wrap_command(cmd)
-
-            result = run(cmd, hide=True, warn=True)
-
-            if result.ok:
-                # Log ssh-copy-id informational output
-                if result.stdout.strip():
-                    for line in result.stdout.strip().splitlines():
-                        self.logger.info(f"ssh-copy-id: {line}")
-
-                # verify we can ssh without password
-                ssh_success = self._verify_ssh_connection(hostname, ssh_conf_path)
-
-                if ssh_success:
-                    return True
-            else:
-                # Log ssh-copy-id output — warning for retries, error on last attempt
-                is_last = attempt == max_retries
-                log_fn = self.logger.error if is_last else self.logger.warning
-                combined = (result.stderr.strip() or result.stdout.strip())
-                if combined:
-                    for line in combined.splitlines():
-                        line = line.strip()
-                        if not line:
-                            continue
-                        log_fn(f"ssh-copy-id: {line}")
-                else:
-                    log_fn("ssh key addition failed (no output)")
-
-            # wait before next attempt with exponential backoff
-            time.sleep(wait_time)
-            wait_time = min(wait_time * 2, max_wait)
-
-        return False
-
-    def _verify_ssh_connection(self, hostname: str, ssh_config_path: str) -> bool:
-        """
-        Verify SSH connection to a VM.
-
-        Args:
-            hostname: Hostname of the VM
-            ssh_config_path: Path to the SSH config file
-
-        Returns:
-            bool: True if successful, False otherwise
-        """
-        self.logger.info(f"verifying ssh connection to: {hostname}")
-        ssh_cmd = f'ssh -o BatchMode=yes -F {ssh_config_path} {hostname} hostname'
-
-        # When using a non-local runtime, the VM is only reachable from
-        # inside the container. The shared ssh_config contains a
-        # ProxyJump directive aimed at the host-side published SSH port
-        # (e.g. 127.0.0.1:2417), which is *not* reachable from inside
-        # the container's own netns — there, 127.0.0.1:2417 is a dead
-        # port. Override ProxyJump/ProxyCommand to "none" so the
-        # in-container ssh hits the VM IP directly (via the `Hostname`
-        # directive in the VM's Host block).
-        if self._runtime_name != 'local':
-            ssh_cmd = (
-                f'ssh -o BatchMode=yes -o ProxyJump=none '
-                f'-o ProxyCommand=none -F {ssh_config_path} '
-                f'{hostname} hostname'
-            )
-        ssh_cmd = self.runtime_instance.wrap_command(ssh_cmd)
-
-        result = run(ssh_cmd, hide=True, warn=True)
-
-        if result.ok and result.stdout.strip():
-            hostname_output = result.stdout.strip()
-            self.logger.info(f"ssh connection verified: {hostname_output}")
-            return True
-        else:
-            self.logger.error(f"ssh connection failed for {hostname}: {result.stderr.strip()}")
-            return False
-
-    def setup_ssh_access(self) -> bool:
-        """
-        Set up SSH access to all VMs.
-
-        This method:
-        1. Writes global authorized keys file (resolved from boxman.yml)
-        2. Generates SSH keys if they don't exist
-        3. Adds the public key to all vms
-        4. Writes an SSH config file for easy access
-
-        Returns:
-            bool: True if all steps completed successfully, False otherwise
-        """
-        # write global authorized keys so they can be consumed by container
-        # entrypoints or cloud-init scripts
-        for _, cluster in self._vm_clusters.items():
-            workdir = os.path.expanduser(cluster['workdir'])
-            global_keys_path = os.path.join(workdir, 'global_authorized_keys')
-            self.write_global_authorized_keys_file(global_keys_path)
-
-        if not self.generate_ssh_keys():
-            self.logger.error("failed to generate ssh keys")
-            return False
-
-        self.write_ssh_config()
-
-        if not self.add_ssh_keys_to_vms():
-            self.logger.error("failed to add ssh keys to some vms")
-            return False
-
-        self.logger.info("")
-        self.logger.info("ssh access setup complete")
-        self.logger.info("you can now connect to vms using the ssh config file")
-
-        return True
 
     def _get_project_vm_names(self) -> list[str]:
         """
