+++ src/boxman/manager_parts/__init__.py
@@ -1,0 +1,1 @@
+"""Mixin modules that make up :class:`boxman.manager.BoxmanManager`."""
